requirements.txt
grobid_client_python==0.0.8
langchain==0.2.8
langchain_community==0.2.7
langchain_core==0.2.19
langchain_experimental==0.0.62
langchain_pinecone==0.1.1
langsmith==0.1.86
pinecone==4.0.0
protobuf==5.27.2
streamlit==1.31.1
streamlit_chat==0.1.1
streamlit_extras==0.4.3
